Update(댓글): 유저인증 관련 #5
유저인증관련데이터를 받고 응답하는
Controller

+++ src/main/java/project_1st_team03/dashboard/domain/comment/application/CommentService.java
@@ -3,14 +3,17 @@ package project_1st_team03.dashboard.domain.comment.application;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import project_1st_team03.dashboard.domain.comment.dao.CommentRepository;
 import project_1st_team03.dashboard.domain.comment.domain.Comment;
 import project_1st_team03.dashboard.domain.comment.dto.CommentsRequest;
 import project_1st_team03.dashboard.domain.comment.dto.CommentsResponse;
+import project_1st_team03.dashboard.domain.comment.exception.CommentException;
 import project_1st_team03.dashboard.domain.member.dao.MemberRepository;
 import project_1st_team03.dashboard.domain.member.domain.Member;
 import project_1st_team03.dashboard.domain.post.dao.PostRepository;
-import project_1st_team03.dashboard.domain.post.domain.Post;
+import project_1st_team03.dashboard.global.exception.ErrorCode;
+import project_1st_team03.dashboard.global.security.MemberDetails;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -22,18 +25,19 @@ public final CommentRepository commentRepository;
 public final MemberRepository memberRepository;
 public final PostRepository postRepository;
 
-    public void createComments(CommentsRequest commentsRequest) {
+    public void createComments(MemberDetails memberDetails,
+            CommentsRequest commentsRequest) {
 
         //Post post = postRepository.findById(commentsRequest.getPostId())
         //        .orElseThrow(() -> new EntityNotFoundException("Post not found"));
 
-      //  long author = Long.parseLong( commentsRequest.getAuthor());
-       // Member member = memberRepository.findById(author)
-        //        .orElseThrow(() -> new EntityNotFoundException("Member not found"));
+        String author =  memberDetails.getUsername();
+        Member member = memberRepository.findByEmail(author)
+                .orElseThrow(EntityNotFoundException::new);
 
         Comment comment = Comment.builder()
                 .content(commentsRequest.getContent())
-                .member(null)
+                .member(member)
                 .post(null)
                 .build();
         commentRepository.save(comment);
@@ -46,7 +50,7 @@ public final PostRepository postRepository;
             CommentsResponse.builder()
                     .id(comment.getId())
                     .content(comment.getContent())
-                    .memberId(1L)
+                    .memberId(comment.getMember().getId())
                     .postId(1L)
                     .createdAt(comment.getCreatedAt())
                     .modifiedAt(comment.getModifiedAt())
@@ -56,13 +60,26 @@ public final PostRepository postRepository;
 
     }
 
-    public void deleteComment(Long commentId) {
+    @Transactional
+    public void deleteComment(MemberDetails memberDetails,Long commentId) {
+        Comment comment = commentRepository.findById(commentId)
+                .orElseThrow(() -> new CommentException(ErrorCode.NOT_FOUND_COMMENT));
+
+        if (!comment.getMember().getEmail().equals(memberDetails.getUsername())) {
+            throw new CommentException(ErrorCode.UNAUTHORIZED_EDIT_COMMENT_ATTEMPT);
+        }
         commentRepository.deleteById(commentId);
     }
 
-    public void updateComment(Long id, CommentsRequest commentRequest) {
+    @Transactional
+    public void updateComment(MemberDetails memberDetails,
+                              Long id, CommentsRequest commentRequest) {
         Comment comment = commentRepository.findById(id).orElseThrow(
-                ()-> new RuntimeException("Comment not found"));
+                ()-> new CommentException(ErrorCode.NOT_FOUND_COMMENT));
+
+        if (!comment.getMember().getEmail().equals(memberDetails.getUsername())) {
+            throw new  CommentException(ErrorCode.UNAUTHORIZED_EDIT_COMMENT_ATTEMPT);
+        }
         comment.setContent(commentRequest.getContent());
         commentRepository.save(comment);
